+++ src/index.ts
@@ -17,7 +17,7 @@ import { AddresController } from './controllers/AddressController'
 import AnswerController from './controllers/AnswerController'
 import AuthController from './controllers/AuthController'
 import BugReportController from './controllers/BugReportController'
-import EmployeeController from './controllers/EmployeeController'
+import EmployeeController from './controllers/employees/EmployeeController'
 import EventController from './controllers/EventController'
 import FileController from './controllers/FileController'
 import UserController from './controllers/UserController'
